feat: refactored funcs and names for initi
